requirements.txt
pandas==2.3.3
openai==2.9.0
streamlit==1.52.1
python-dotenv==1.2.1
